refact(StringCalculator) : divide responsibility
