+++ src/test/java/calculator/PositiveNumber.java
@@ -1,0 +1,16 @@
+package calculator;
+
+public class PositiveNumber {
+	private final int number;
+
+	public PositiveNumber(String text) {
+		if (Integer.parseInt(text) < 0) {
+			throw new RuntimeException();
+		}
+		number = Integer.parseInt(text);
+	}
+
+	public int getNumber() {
+		return number;
+	}
+}
