+++ src/test/java/calculator/StringCalculatorTest.java
@@ -47,7 +47,7 @@ public class StringCalculatorTest {
 
 	@DisplayName(value = "//와 \\n 문자 사이에 커스텀 구분자를 지정할 수 있다.")
 	@ParameterizedTest
-	@ValueSource(strings = {"//;\n1;2;3"})
+	@ValueSource(strings = {"//;\n1;2,3"})
 	void test5(final String text) {
 		assertThat(calculator.add(text)).isSameAs(6);
 	}
